requirements.txt
deepl-translate
streamlit